create GeneralMetatags component
